This adds Status.WatchForUpdate().
This allows a client to be notified when part of the status structure
is updated, based on a URL. The URL can contain wildcards of the form:

  status://foo/*/bar/

The URL can refer to an explicit value, or to a sub-tree. Or even the
whole tree, if the URL status:// is used.