Avoid deadlock if the clients never read.
If there is a pending notification that the client hasn't read, it
would block our wite (bigger buffers just delay this). So, we use a
buffer of one, and clear it before writing to it.

The client is now garanteed to be notified if there is a change since
they checked, but not garanteed to be notified of every change.