logging: Add a new cached logging mechanism.
Add CachedLogging, which can store the 100 most recent log lines
in memory, and notify interested listeners when new logs are added.

Add LogHandler to the http-server, and add support in /log/ for
serving log entries to interested clients in Json format.

Wire it all up in go-house, but keep logging to Stderr as well.