sparkAdapter: Remeber events after device update.
If a device is updated (which can happen for reasons unrelated to the
device), preserve any events we've received from it.

This keeps us from losing event data, but DOES mean that we'll never
ever drop an event if the core software stops sending it until there
is a server restart.

Also, I found there was no adaptor event testing (bad me!), so I
created some.