wait: Test helper method wait.Wait.
During test for async results, we want to finish as soon as possible
while having a long enough timeout to make a test robust. One way to
do this is test repeatedly to see if the end condition has been met
(and exit, if so).

This wait method gives a general implementation of that wait loop.