conditions.base: Create sendResult method.
All conditions now use a helper method to send results to
listeners. This helper avoids sending duplicate updates, and ensures
that we never block when sending them.

If a client isn't reading updates, some updates can be lost, but the
most recent 2 updates will always be queued. Since we are sending
boolean values, that means at least one of each value will always be
ready to be received.