Add new Time condition.
This condition is intended to be a more general version of daily and
periodic, and will replace both of them after further testing.

This new condition is a fairly thin wrapper around the timeglob
library, and is the first real user of that library, so all kinds of
bugs might shake out.